+++ structurizr-core/src/com/structurizr/model/Relationship.java
@@ -1,6 +1,7 @@
 package com.structurizr.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.structurizr.util.Url;
 
 import java.util.Collections;
 import java.util.LinkedHashSet;
@@ -18,6 +19,7 @@ public final class Relationship extends ModelItem {
     private String description;
     private String technology;
     private InteractionStyle interactionStyle = InteractionStyle.Synchronous;
+    private String url;
 
     private String linkedRelationshipId;
 
@@ -142,6 +144,31 @@ public final class Relationship extends ModelItem {
             return new LinkedHashSet<>(Collections.singletonList(Tags.RELATIONSHIP));
         } else {
             return Collections.emptySet();
+        }
+    }
+
+    /**
+     * Gets the URL where more information about this relationship can be found.
+     *
+     * @return  a URL as a String
+     */
+    public String getUrl() {
+        return url;
+    }
+
+    /**
+     * Sets the URL where more information about this relationship can be found.
+     *
+     * @param url   the URL as a String
+     * @throws IllegalArgumentException     if the URL is not a well-formed URL
+     */
+    public void setUrl(String url) {
+        if (url != null && url.trim().length() > 0) {
+            if (Url.isUrl(url)) {
+                this.url = url;
+            } else {
+                throw new IllegalArgumentException(url + " is not a valid URL.");
+            }
         }
     }