+++ docs/changelog.md
@@ -7,6 +7,7 @@
 - Adds support for themes.
 - Adds support for tags on deployment nodes.
 - Adds support for animations on deployment views.
+- Adds support for URLs on relationships.
 
 ## 1.3.3 (24th December 2019)